+++ src/app/brokers/alpaca/dashboard/page.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from 'next-intl';
 import AlpacaHistoryChart from '@/components/alpaca/history-charts/AlpacaHistoryChart';
 import * as React from 'react';
-import "../../globals.css";
+import "@/app/globals.css";
 import AlpacaAssets from '@/components/alpaca/assets/AlpacaAssets';