+++ src/app/brokers/layout.tsx
@@ -1,6 +1,6 @@
 import BrokerNavi from "@/components/BrokerNavi";
 import Link from "next/link";
-import "../../app/globals.css";
+import "@/app/globals.css";
 
 const BlogEintries = ['2024-8-1', '2024-8-2', '2024-8-3', '2024-8-4'];
 
@@ -17,10 +17,8 @@ export default async function BrokerLayout({
         <h6 className="mb-4">Aktuell verfügbare Broker</h6>
         <BrokerNavi />
       </nav>
-      <main>
-        <div className="h-full w-full content-container">
-          {children}
-        </div>
+      <main className="h-full w-full content-container">
+        {children}
       </main>
     </div>
   );