+++ src/model/AlpacaAccount.ts
@@ -1,0 +1,7 @@
+export interface AccountData  {
+    accountId: string
+    accountNumber: string
+    accruedFees: number
+    buyingPower: number
+    createdAtUtc: Date
+}