+++ src/service/alpacaClientService.ts
@@ -2,8 +2,6 @@ import fetchWithCache, { basicFetch , basicPost} from "../app/api/fetchFunctions
 
 export const alpacaClientService = {
 
-  baseURL: process.env.BASE_API_URL,
-
   async toggleAssetSelection(symbol: string) {
     try {
       const url = `/api/alpaca/assets`;